requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
